+++ components/atoms/Pill/pill.tsx
@@ -21,7 +21,7 @@ const Pill: React.FC<PillProps> = ({ className, text, color = "slate", size = "b
             ? "bg-light-red-4 "
             : color === "purple"
             ? "bg-purple-200"
-            : "bg-light-slate-4 "
+            : "bg-light-slate-3 "
         }
         ${size === "small" ? "py-1 px-1.5  " : "py-1.5 px-2 "}
         inline-flex items-center rounded-full gap-1 ${className}`}